# openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "demo API (generated from code)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/admin/complaints": {
      "get": {
        "tags": [
          "管理员-投诉管理"
        ],
        "summary": "查看所有投诉",
        "description": "管理员查看平台所有投诉记录",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 1,
                      "msg": "success",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "JWT token（项目使用请求头 token）"
          }
        ]
      }
    },
    "/admin/complaints/stats": {
      "get": {
        "tags": [
          "管理员-投诉管理"
        ],
        "summary": "统计投诉状态",
        "description": "统计不同状态的投诉数量",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 1,
                      "msg": "success",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "JWT token（项目使用请求头 token）"
          }
        ]
      }
    },
    "/admin/complaints/{id}/resolve": {
      "post": {
        "tags": [
          "管理员-投诉管理"
        ],
        "summary": "处理投诉",
        "description": "管理员处理用户投诉",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "examples": {
                  "success": {
                    "value": {
                      "code": 1,
                      "msg": "success",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "JWT token（项目使用请求头 token）"
          }
        ]
      }
    },
...
  }
}